# users-service/src/main/gen/api/openapi.yaml
openapi: 3.0.1
info:
  description: API para gestionar los usuarios del sistema
  title: User-service API
  version: 1.0.0
servers:
- url: http://localhost:8080/api
paths:
  /artist/public/{artistUserName}:
    get:
      parameters:
      - description: Nombre de usuario del artista.
        explode: false
        in: path
        name: artistUserName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SentArtistDTO'
          description: Informacion pública del artista recuperada con éxito
        "404":
          content:
            application/json:
              example:
                error: Not Found
                message: Artista no encontrado
                statusCode: 404
                timestamp: 2025-10-01T12:03:00Z
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "Artista {artistUserName} no encontrado"
      summary: Recupera la información pública de un artista a partir de su nombre
        de usuario.
      tags:
      - Artistas
      x-accepts:
      - application/json
  /artist/public:
    get:
      description: Recupera una página de artistas filtrados opcionalmente por nombre
        y/o género musical. Soporta paginación mediante los parámetros `page` y `size`.
      parameters:
      - description: Nombre parcial del artista para filtrar.
        example: adam
        explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      - description: Conjunto de géneros para filtrar artistas
        example:
        - HIP_HOP
        - RAP
        explode: false
        in: query
        name: genres
        required: false
        schema:
          items:
            $ref: '#/components/schemas/GenreType'
          type: array
        style: form
      - description: Número de páginas solicitadas (empezando en 0)
        example: 0
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - description: Tamaño de la página
        example: 10
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSentArtistDTO'
          description: Página de artistas coincidentes recuperadas con éxito
        "404":
          content:
            application/json:
              example:
                error: Not Found
                message: No hay artistas que coinciden con la selección de búsqueda
                statusCode: 404
                timestamp: 2025-10-01T12:02:00Z
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: No hay artistas que coinciden con la selección de búsqueda
      summary: Recupera una página de artistas filtrados
      tags:
      - Artistas
      x-accepts:
      - application/json
  /artist/payment-info:
    get:
      description: "Devuelve los balances, pagos y otra información financiera del\
        \ artista autenticado. Requiere autenticación."
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistPayementInfoDTO'
          description: Información financiera del artista recuperada con éxito
        "401":
          content:
            application/json:
              example:
                error: unauthorized
                message: No tienes permisos para obtener la información financiera
                  del artista
                statusCode: 401
                timestamp: 2025-10-01T12:00:00Z
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: No tienes permisos para obtener la información financiera del
            artista
      security:
      - BearerAuth: []
      summary: Obtiene toda la información financiera de un artista autenticado.
      tags:
      - Artistas
      x-accepts:
      - application/json
  /artist/public/trending:
    get:
      description: Devuelve una lista de los artistas más populares marcados como
        "trending".
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SentArtistDTO'
                type: array
          description: Lista de artistas en tendencia recuperada con éxito
        "404":
          content:
            application/json:
              example:
                error: Not Found
                message: No existe por el momento ningún artista en tendencia
                statusCode: 404
                timestamp: 2025-10-01T11:59:00Z
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: No existe por el momento ningún artista en tendencia
      summary: Recupera una lista de los artistas en tendencia
      tags:
      - Artistas
      x-accepts:
      - application/json
  /artist/{id}/restauracion:
    patch:
      description: Restaura a un artista previamente eliminado de forma lógica. Requiere
        que el usuario autenticado sea el dueño de la cuenta de artista.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          example: 4
          type: number
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulResponseDTO'
          description: Artista restaurado correctamente
        "400":
          description: El artista no está eliminado
        "401":
          content:
            application/json:
              example:
                error: unauthorized
                message: No tienes permitido restaurar a un artista
                statusCode: 401
                timestamp: 2025-10-01T14:41:00
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: No está permitido que restaures a un artista
      security:
      - BearerAuth: []
      summary: Restaura a un artista previamente eliminado de forma lógica
      tags:
      - Artistas
      x-accepts:
      - application/json
  /artist:
    delete:
      description: Realiza un borrado lógico de un artista autenticado junto con toda
        su información.El artista y sus canciones/álbumes aparecen borrados pero no
        se han eliminado de manera física.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulResponseDTO'
          description: Artista borrado de forma existosa
        "400":
          content:
            application/json:
              example:
                error: Bad Request
                message: Artista ya borrado o operación inválida
                statusCode: 400
                timestamp: 2025-10-01T11:55:00Z
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Mala petición - Artista ya borrado o operación inválida
        "401":
          content:
            application/json:
              example:
                error: unauthorized
                message: No tienes permitido borrar a un artista
                statusCode: 401
                timestamp: 2025-10-01T14:41:00
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Sin autorizacion - No tienes permisos para borrar a un artista
      security:
      - BearerAuth: []
      summary: Realiza un borrado lógico de un artista autenticado
      tags:
      - Artistas
      x-accepts:
      - application/json
    patch:
      description: Permite que un artista autenticado actualice su información.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulResponseDTO'
          description: Artista actualizado correctamente
        "401":
          content:
            application/json:
              example:
                error: unauthorized
                message: No tienes permitido actualizar a un artista
                statusCode: 401
                timestamp: 2025-10-01T14:41:00
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: No tienes permisos para actualizar a un artista
      security:
      - BearerAuth: []
      summary: Actualiza los datos de un artista autenticado
      tags:
      - Artistas
      x-content-type: application/json
      x-accepts:
      - application/json
    post:
      description: Convierte un usuario autentificado en un artista con datos incluidos.
        Si el usuario ya está registrado como artista se manejara un error de contenido.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistFormDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulResponseDTO'
          description: El usuario ha sido existosamente convertido a artista
        "401":
          content:
            application/json:
              example:
                error: unauthorized
                message: No tienes permitido convertir un usuario a artista
                statusCode: 401
                timestamp: 2025-10-01T14:41:00
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Sin autorizacion - No tienes permisos para convertir un usuario
            a artista
      security:
      - BearerAuth: []
      summary: Convierte un usuario en artista
      tags:
      - Artistas
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
...
    ArtistFormDTO:
      example:
        iban: ES7921000813610123456789
        description: WILYWNKA es un cantante de género RAP|J_POP ....
        artistName: WILYWNKA
        accountPropietary: Company EPSILON
      properties:
        artistName:
          description: El nombre del artista debe tener entre 2 y 100 caracteres
          example: WILYWNKA
          maxLength: 100
          minLength: 2
          type: string
        iban:
          description: El iban no debe estar vacío
          example: ES7921000813610123456789
          pattern: "^[A-Z]{2}\\d{2}[A-Z0-9]{1,30}$"
          type: string
        accountPropietary:
          description: El propietario de la cuenta debe tener entre 2 y 100 caracteres
          example: Company EPSILON
          maxLength: 100
          minLength: 2
          type: string
        description:
          description: La descripción no debe de ser vacía
          example: WILYWNKA es un cantante de género RAP|J_POP ....
          maxLength: 1000
          minLength: 10
          type: string
      required:
      - accountPropietary
      - artistName
      - description
      - iban
      type: object
    SuccessfulResponseDTO:
      example:
        message: El artista ha sido actualizado correctamente
        successful: successful_artist_update
        statusCode: 200
        timestamp: 01/10/2025 13:13:00
      properties:
        successful:
          example: successful_artist_update
          type: string
        message:
          example: El artista ha sido actualizado correctamente
          type: string
        statusCode:
          example: 200
          type: integer
        timestamp:
          example: 01/10/2025 13:13:00
          type: string
      type: object
    ErrorResponse:
      example:
        error: unauthorized
        message: Debes iniciar sesión para poder acceder a este recurso
        statusCode: 401
        timestamp: 2025-10-01T16:30:00Z
      properties:
        error:
          description: Código corto de error
          example: unauthorized
          type: string
        message:
          description: Mensaje descriptivo del error
          example: Debes iniciar sesión para poder acceder a este recurso
          type: string
        statusCode:
          description: Código HTTP asociado al error
          example: 401
          type: integer
        timestamp:
          description: Hora en la que se ha producido el error
          example: 2025-10-01T16:30:00Z
          format: date-time
          type: string
      type: object
    UpdateArtistDTO:
      example:
        artisticName: Mathew PJ
        genres:
        - ROCK
        - JAZZ
        iban: ES7921000813610123456789
        description: Músico especializado en jazz y soul.
        socialMediaLinks:
        - id: 1
          type: SPOTIFY
        - id: 2
          type: INSTAGRAM
      properties:
        artisticName:
          description: El nombre artístico no debe estar en blanco (entre 2 y 100
            caracteres)
          example: Mathew PJ
          maxLength: 100
          minLength: 2
          type: string
        description:
          description: La describición no debe estar en blanco (entre 10 y 1000 caracteres)
          example: Músico especializado en jazz y soul.
          maxLength: 1000
          minLength: 10
          type: string
        iban:
          description: El formato del iban debe de ser válido
          example: ES7921000813610123456789
          pattern: "^[A-Z]{2}\\d{2}[A-Z0-9]{1,30}$"
          type: string
        genres:
          description: Debes proprocionar entre 1 y 4 géneros distintos
          example:
          - ROCK
          - JAZZ
          items:
            $ref: '#/components/schemas/GenreType'
          maxItems: 4
          minItems: 1
          type: array
        socialMediaLinks:
          description: Debes proporcionar entre 1 y 4 enlaces a tus redes sociales.
          example:
          - id: 1
            type: SPOTIFY
          - id: 2
            type: INSTAGRAM
          items:
            $ref: '#/components/schemas/SocialMediaLinks'
          maxItems: 4
          minItems: 1
          type: array
      required:
      - accountProperty
      - artisticName
      - description
      - genres
      - iban
      - socialMediaLinks
      type: object
    SentArtistDTO:
      example:
        artisticName: TORAUMA
        monthlyStreams: 4
        isTrending: true
        genres:
        - id: 4
          type: "J_POP, RAP, HIP_HOP"
        - id: 4
          type: "J_POP, RAP, HIP_HOP"
        verified: true
        description: Cantante inigualable que promueve el género Hip-Hop | Rap moderno
          en Japón
        artistId: 100
        socialMediaLinks:
        - id: 1
          type: SPOTIFY
        - id: 1
          type: SPOTIFY
        username: torauma@111
      properties:
        artistId:
          example: 100
          type: number
        username:
          example: torauma@111
          type: string
        artisticName:
          example: TORAUMA
          type: string
        description:
          example: Cantante inigualable que promueve el género Hip-Hop | Rap moderno
            en Japón
          type: string
        isTrending:
          example: true
          type: boolean
        verified:
          example: true
          type: boolean
        socialMediaLinks:
          items:
            $ref: '#/components/schemas/SocialMediaLinks'
          type: array
        monthlyStreams:
          example: 4
          type: number
        genres:
          items:
            $ref: '#/components/schemas/Genre'
          type: array
      type: object
    ArtistPayementInfoDTO:
      example:
        balance:
        - balanceDate: 30-09-2025
          monthlyBalance: 1030.0
          artistId: 4
          id: 30
          status: PENDIENTE
        - balanceDate: 30-09-2025
          monthlyBalance: 1030.0
          artistId: 4
          id: 30
          status: PENDIENTE
        totalBalance: 1030.0
        iban: ES9121000418450200051332
        thisMonth: 444.4
        accountPropietary: jason the wheen
        username: "4"
        status:
        - amount: 2
          concept: Compra de canción
          paymentMethod:
          - TRANSFERENCIA
          - TRANSFERENCIA
          artistId: 4
          artistName: jason
          id: 1
          paymentDate: 12-09-2025
          status:
          - PENDIENTE
          - PENDIENTE
        - amount: 2
          concept: Compra de canción
          paymentMethod:
          - TRANSFERENCIA
          - TRANSFERENCIA
          artistId: 4
          artistName: jason
          id: 1
          paymentDate: 12-09-2025
          status:
          - PENDIENTE
          - PENDIENTE
      properties:
        username:
          example: "4"
          type: string
        totalBalance:
          example: 1030.0
          type: number
        thisMonth:
          example: 444.4
          type: number
        iban:
          example: ES9121000418450200051332
          type: string
        accountPropietary:
          example: jason the wheen
          type: string
        balance:
          items:
            $ref: '#/components/schemas/BalanceDTO'
          type: array
        status:
          items:
            $ref: '#/components/schemas/PaymentDTO'
          type: array
      type: object
    PaymentDTO:
      example:
        amount: 2
        concept: Compra de canción
        paymentMethod:
        - TRANSFERENCIA
        - TRANSFERENCIA
        artistId: 4
        artistName: jason
        id: 1
        paymentDate: 12-09-2025
        status:
        - PENDIENTE
        - PENDIENTE
      properties:
        id:
          example: 1
          type: number
        artistId:
          example: 4
          type: number
        artistName:
          example: jason
          type: string
        concept:
          example: Compra de canción
          type: string
        paymentDate:
          example: 12-09-2025
          type: string
        amount:
          example: 2
          type: number
        paymentMethod:
          items:
            $ref: '#/components/schemas/PaymentMethod'
          type: array
        status:
          items:
            $ref: '#/components/schemas/PaymentStatus'
          type: array
      type: object
    PaymentStatus:
      enum:
      - PENDIENTE
      - COMPLETADO
      - FALLADO
      type: string
    PaymentMethod:
      enum:
      - TRANSFERENCIA
      - PAYPAL
      - TARJETA_CREDITO
      - EFECTIVO
      - OTRO
      type: string
    BalanceDTO:
      example:
        balanceDate: 30-09-2025
        monthlyBalance: 1030.0
        artistId: 4
        id: 30
        status: PENDIENTE
      properties:
        id:
          example: 30
          type: number
        artistId:
          example: 4
          type: number
        monthlyBalance:
          example: 1030.0
          type: number
        balanceDate:
          example: 30-09-2025
          type: string
        status:
          $ref: '#/components/schemas/BalanceStatus'
      type: object
    BalanceStatus:
      enum:
      - PENDIENTE
      - PAGADO
      - RECHAZADO
      type: string
    PageSentArtistDTO:
      description: Representación de una página de resultados de artistas
      example:
        content:
        - artisticName: TORAUMA
          monthlyStreams: 4
          isTrending: true
          genres:
          - id: 4
            type: "J_POP, RAP, HIP_HOP"
          - id: 4
            type: "J_POP, RAP, HIP_HOP"
          verified: true
          description: Cantante inigualable que promueve el género Hip-Hop | Rap moderno
            en Japón
          artistId: 100
          socialMediaLinks:
          - id: 1
            type: SPOTIFY
          - id: 1
            type: SPOTIFY
          username: torauma@111
        - artisticName: TORAUMA
          monthlyStreams: 4
          isTrending: true
          genres:
          - id: 4
            type: "J_POP, RAP, HIP_HOP"
          - id: 4
            type: "J_POP, RAP, HIP_HOP"
          verified: true
          description: Cantante inigualable que promueve el género Hip-Hop | Rap moderno
            en Japón
          artistId: 100
          socialMediaLinks:
          - id: 1
            type: SPOTIFY
          - id: 1
            type: SPOTIFY
          username: torauma@111
      properties:
        content:
          items:
            $ref: '#/components/schemas/SentArtistDTO'
          type: array
      type: object
    SocialMediaLinks:
      example:
        id: 1
        type: SPOTIFY
      properties:
        id:
          example: 1
          type: number
        type:
          $ref: '#/components/schemas/SocialMediaTypes'
      required:
      - id
      - type
      type: object
    SocialMediaTypes:
      enum:
      - INSTAGRAM
      - SPOTIFY
      - FACEBOOK
      - GITHUB
      - YOUTUBE
      - SOUNDCLOUD
      - CUENTA_UNEX
      type: string
    Genre:
      example:
        id: 4
        type: "J_POP, RAP, HIP_HOP"
      properties:
        id:
          example: 4
          type: number
        type:
          $ref: '#/components/schemas/GenreType'
      required:
      - id
      - type
      type: object
    GenreType:
      enum:
      - HIP_HOP
      - RAP
      - TRAP
      - RNB
      - POP
      - ROCK
      - ALTERNATIVE_ROCK
      - HARD_ROCK
      - PUNK_ROCK
      - METAL
      - HEAVY_METAL
      - TRASH_METAL
      - INDIE
      - INDIE_ROCK
      - INDIE_POP
      - ELECTRONIC
      - EDM
      - HOUSE
      - TECHNO
      - TRANCE
      - DUBSTEP
      - DRUM_AND_BASS
      - JAZZ
      - BLUES
      - FUNK
      - SOUL
      - CLASSICAL
      - REGGAE
      - REGGAETON
      - SALSA
      - BACHATA
      - CUMBIA
      - FLAMENCO
      - FOLK
      - COUNTRY
      - LATIN
      - AFROBEAT
      - K_POP
      - J_POP
      - LOFI
      - AMBIENT
      - EXPERIMENTAL
      - INSTRUMENTAL
      - ACOUSTIC
      - SOUNDTRACK
      - GOSPEL
      - OPERA
      - DANCEHALL
      - DISCO
      - GRIME
      - DRILL
      - PHONK
      - CHILLWAVE
      - SYNTHWAVE
      - NEW_WAVE
      type: string
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X_API_Key
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http